SERVER: Change transaction processing logic

+++ commons/src/main/java/org/company/model/Player.java
@@ -1,11 +1,12 @@
 package org.company.model;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 
 public record Player(
     String username,
-    long balanceVersion,
+    BigInteger balanceVersion,
     BigDecimal balance
 ) {
 }
